wip - refactor entry_types to separate files
