+++ src/entry_type.rs
@@ -1,354 +1,12 @@
-use std::path::{Path, PathBuf};
-use std::fmt;
-use std::collections::HashMap;
+use std::path::Path;
 use crate::models::SessionId;
+use crate::models::Session;
+use crate::models::MovieName;
+use crate::models::EntryType;
+use crate::models::EncodeDirType;
+use crate::models::TVSeriesEncodeDir;
+use crate::models::MovieEncodeDir;
 
-// TODO: Extend EntryType to have one rename for tv series and one for movies
-// This class models something that can be encoded later.
-#[derive(Debug, Clone)]
-pub enum EntryType {
-  TVSeriesRename {
-
-    /// Full path to mkv file
-    path: PathBuf,
-
-    /// Session id of file
-    session: SessionId,
-
-    /// Episode
-    episode: String,
-
-    /// file name and extension
-    file: String
-  },
-
-  MovieRename {
-
-    /// Full path to mkv file
-    path: PathBuf,
-
-    /// Session id of file
-    session: SessionId,
-
-    /// file name and extension
-    file: String
-  },
-
-  TVSeriesEncode {
-
-    /// Session id of files that map to this encode directory
-    session: SessionId,
-
-    /// Full path to encode directory
-    path: PathBuf,
-
-    // TODO: What does this mean in terms of a movie? Is it the name? Choose a better name for this field
-    /// Season
-    season: String,
-  },
-
-  MovieEncode {
-
-    /// Session id of files that map to this encode directory
-    session: SessionId,
-
-    /// Full path to encode directory
-    path: PathBuf,
-
-    /// Name of the movie
-    movie_name: MovieName
-  },
-
-  UnknownFileType {
-    path: PathBuf
-  },
-
-
-  InvalidEncodeDirPath {
-    defined_path: String
-  }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Ord, PartialOrd)]
-pub struct MovieName(String);
-
-impl MovieName {
-  pub fn new(value: &str) -> Self {
-    Self(value.to_string())
-  }
-
-  pub fn id(&self) -> &str {
-    &self.0
-  }
-}
-
-impl fmt::Display for MovieName {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-      write!(f, "{}", self.0)
-  }
-}
-
-#[derive(Debug, Clone)]
-pub enum RenameTypes {
-  TVSeries(TVSeriesRenameFile),
-  Movie(MovieRenameFile)
-}
-
-
-#[derive(Debug, Clone)]
-pub struct TVSeriesRenameFile {
-
-  /// Full path to mkv_file
-  pub path: PathBuf,
-
-  /// Session id associated with file
-  pub session: SessionId,
-
-  /// Episode name
-  pub episode: String,
-
-  /// Input file name and ext - file to be encoded
-  pub mkv_file: String,
-
-  /// Output file and ext - encoded file
-  pub mp4_file: String,
-}
-
-pub trait SessionTypeAware {
-  fn session_id(&self) -> SessionId;
-}
-
-pub trait EpisodeName {
-  fn episode(&self) -> Option<String>;
-}
-
-pub trait MKVTypeAware {
-  fn mkv_file(&self) -> String;
-  fn mp4_file(&self) -> String;
-  fn mkv_path(&self) -> PathBuf;
-}
-
-
-impl SessionTypeAware for RenameTypes {
-  fn session_id(&self) -> SessionId {
-    match self {
-        RenameTypes::TVSeries(tvseries_rename_file) => tvseries_rename_file.session.clone(),
-        RenameTypes::Movie(movie_rename_file) => movie_rename_file.session.clone(),
-    }
-  }
-}
-
-impl MKVTypeAware for RenameTypes {
-  fn mkv_file(&self) -> String {
-    match self {
-      RenameTypes::TVSeries(tvseries_rename_file) => tvseries_rename_file.mkv_file.to_owned(),
-      RenameTypes::Movie(movie_rename_file) => movie_rename_file.mkv_file.to_owned(),
-    }
-  }
-
-  fn mp4_file(&self) -> String {
-    match self {
-      RenameTypes::TVSeries(tvseries_rename_file) => tvseries_rename_file.mp4_file.to_owned(),
-      RenameTypes::Movie(movie_rename_file) => movie_rename_file.mp4_file.to_owned(),
-    }
-  }
-
-  fn mkv_path(&self) -> PathBuf {
-    match self {
-      RenameTypes::TVSeries(tvseries_rename_file) => tvseries_rename_file.path.clone(),
-      RenameTypes::Movie(movie_rename_file) => movie_rename_file.path.clone(),
-    }
-  }
-}
-
-impl EpisodeName for RenameTypes {
-    fn episode(&self) -> Option<String> {
-      match self {
-        RenameTypes::TVSeries(tvseries_rename_file) => Some(tvseries_rename_file.episode.to_string()),
-        RenameTypes::Movie(_) => None,
-      }
-    }
-}
-
-/// Convert from a collection of RenameFile into a Map<SessionId, Session>
-impl FromIterator<RenameTypes> for HashMap<SessionId, Session> {
-
-  fn from_iter<T: IntoIterator<Item = RenameTypes>>(renames: T) -> Self {
-    let mut hash: HashMap<SessionId, Vec<RenameTypes>> = HashMap::new();
-      for rename in renames {
-        hash
-          .entry(rename.session_id())
-          .and_modify(|v| v.push(rename.clone()))
-          .or_insert(vec![rename]);
-      }
-
-      hash
-        .into_iter()
-        .map(|(k, v)| (k.clone(), Session::new(k, v)))
-        .collect()
-    }
-}
-
-impl TryFrom<EntryType> for RenameTypes {
-  type Error = ();
-
-  fn try_from(value: EntryType) -> Result<Self, Self::Error> {
-    match value {
-      EntryType::TVSeriesRename { path, session, episode, file } => {
-        let output_path = Path::new(&file);
-
-        let mp4_file =
-          output_path
-            .file_stem()
-            .map(|f| format!("{}.mp4", f.to_string_lossy()))
-            .expect("Could not get file stem");
-
-        let mkv_file = file;
-
-        Ok(
-          RenameTypes::TVSeries(TVSeriesRenameFile {
-            path,
-            session,
-            episode,
-            mkv_file,
-            mp4_file,
-          }
-        ))
-      },
-      EntryType::MovieRename { path, session, file } => {
-        let output_path = Path::new(&file);
-
-        let mp4_file =
-          output_path
-            .file_stem()
-            .map(|f| format!("{}.mp4", f.to_string_lossy()))
-            .expect("Could not get file stem");
-
-        let mkv_file = file;
-
-        Ok(
-          RenameTypes::Movie(MovieRenameFile {
-            path,
-            session,
-            mkv_file,
-            mp4_file,
-          }
-        ))
-      },
-      _ => Err(()),
-    }
-  }
-}
-
-
-#[derive(Debug, Clone)]
-pub struct MovieRenameFile {
-
-  /// Full path to mkv_file
-  pub path: PathBuf,
-
-  /// Session id associated with file
-  pub session: SessionId,
-
-  /// Input file name and ext - file to be encoded
-  pub mkv_file: String,
-
-  /// Output file and ext - encoded file
-  pub mp4_file: String,
-}
-
-
-#[derive(Debug, Clone)]
-pub enum EncodeDirType {
-  TVSeries(TVSeriesEncodeDir),
-  Movie(MovieEncodeDir),
-}
-
-pub trait EncodeDirPathAware {
-  fn path(&self) -> PathBuf;
-}
-
-pub trait LocationAware {
-  fn location(&self) -> String;
-}
-
-#[derive(Debug, Clone)]
-pub struct TVSeriesEncodeDir {
-  pub path: PathBuf,
-  pub season: String,
-  pub session_id: SessionId,
-}
-
-#[derive(Debug, Clone)]
-pub struct MovieEncodeDir {
-  pub path: PathBuf,
-  pub session_id: SessionId,
-  pub movie_name: MovieName
-}
-
-impl TryFrom<EntryType> for EncodeDirType {
-    type Error = ();
-
-    fn try_from(value: EntryType) -> Result<Self, Self::Error> {
-      match value {
-        EntryType::TVSeriesEncode { path, season, session } => {
-          let session_id = session;
-          Ok(
-            EncodeDirType::TVSeries(
-              TVSeriesEncodeDir {
-                path,
-                season,
-                session_id
-              }
-            )
-          )
-        },
-        EntryType::MovieEncode { path, session, movie_name } => {
-          let session_id = session;
-          Ok(
-            EncodeDirType::Movie(
-              MovieEncodeDir {
-                path,
-                session_id,
-                movie_name
-              }
-            )
-          )
-        },
-        _ => {
-          Err(())
-        },
-    }
-  }
-}
-
-impl SessionTypeAware for EncodeDirType {
-    fn session_id(&self) -> SessionId {
-      match self {
-        EncodeDirType::TVSeries(tvseries_encode_dir) => tvseries_encode_dir.session_id.clone(),
-        EncodeDirType::Movie(movie_encode_dir) => movie_encode_dir.session_id.clone(),
-      }
-    }
-}
-
-impl EncodeDirPathAware for EncodeDirType {
-    fn path(&self) -> PathBuf {
-        match self {
-            EncodeDirType::TVSeries(tvseries_encode_dir) => tvseries_encode_dir.path.clone(),
-            EncodeDirType::Movie(movie_encode_dir) => movie_encode_dir.path.clone(),
-        }
-    }
-}
-
-
-impl LocationAware for EncodeDirType {
-    fn location(&self) -> String {
-      match self {
-        EncodeDirType::TVSeries(tvseries_encode_dir) => tvseries_encode_dir.season.clone(),
-        EncodeDirType::Movie(movie_encode_dir) => movie_encode_dir.movie_name.to_string(),
-      }
-    }
-}
 
 impl EntryType {
   pub fn new_tv_series_rename<P: AsRef<Path>>(path: P, session: &str, episode: &str, file: &str) -> Self {
@@ -397,29 +55,6 @@ impl EntryType {
   }
 }
 
-/// A Session has a SessionId and a list of files (RenameFile)
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-pub struct Session {
-  session_id: SessionId,
-  files: Vec<RenameTypes>,
-}
-
-impl Session {
-
-  pub fn new(session_id: SessionId, files: Vec<RenameTypes>) -> Self {
-    Self {
-      session_id,
-      files
-    }
-  }
-
-  pub fn files(&self) -> Vec<RenameTypes> {
-    let mut sorted_files = self.files.clone();
-    sorted_files.sort_by_key(|a| a.mkv_file());
-    sorted_files
-  }
-}
 
 #[derive(Debug)]
 pub struct SessionToEncodeDir {
