+++ src/file_mapper.rs
@@ -2,10 +2,15 @@ use std::collections::HashMap;
 use std::path::Path;
 use walkdir::WalkDir;
 use regex::Regex;
-use crate::{debug::*, entry_type::{RenameTypes, SessionTypeAware}};
+use crate::{debug::*};
 use std::sync::LazyLock;
-use crate::entry_type::{EncodeDirType, EntryType, Session, SessionToEncodeDir};
+use crate::entry_type::SessionToEncodeDir;
 use crate::models::SessionId;
+use crate::models::RenameTypes;
+use crate::models::SessionTypeAware;
+use crate::models::Session;
+use crate::models::EntryType;
+use crate::models::EncodeDirType;
 
 static RENAME_TV_SERIES_FILE_REG: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(session\d{1,})\/renames\/((S\d{2,}E\d{2,})\s-\s(.+.mkv))$").unwrap());
 static RENAME_MOVIE_FILE_REG: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(session\d{1,})\/renames\/(.+.mkv)$").unwrap());
