+++ src/models.rs
@@ -1,3 +1,20 @@
 mod session_id;
+mod movie_name;
+mod rename_types;
+mod session;
+mod entry_type;
+mod encode_dir_type;
 
 pub use session_id::SessionId as SessionId;
+pub use movie_name::MovieName as MovieName;
+pub use rename_types::RenameTypes as RenameTypes;
+pub use rename_types::SessionTypeAware as SessionTypeAware;
+pub use rename_types::EpisodeName as EpisodeName;
+pub use rename_types::MKVTypeAware as MKVTypeAware;
+pub use session::Session as Session;
+pub use entry_type::EntryType as EntryType;
+pub use encode_dir_type::EncodeDirType as EncodeDirType;
+pub use encode_dir_type::EncodeDirPathAware as EncodeDirPathAware;
+pub use encode_dir_type::LocationAware as LocationAware;
+pub use encode_dir_type::TVSeriesEncodeDir as TVSeriesEncodeDir;
+pub use encode_dir_type::MovieEncodeDir as MovieEncodeDir;
