+++ src/user_selection.rs
@@ -1,8 +1,10 @@
 use std::fmt;
 use console::style;
 
-use crate::{entry_type::{EncodeDirType, RenameTypes, SessionToEncodeDir}, profiles::ProfileConfigItem};
+use crate::{entry_type::SessionToEncodeDir, profiles::ProfileConfigItem};
 use crate::models::SessionId;
+use crate::models::RenameTypes;
+use crate::models::EncodeDirType;
 
 pub struct UserSelection {
   session_id: SessionId,
